fix BaseProvider.server_settings
